Moved all the components of the game into a seperate package for all furuter card games.
Split the Player class into a parent and child for inheritence.
